refactor:  remove BaseBadge. Now only Badge component exists
