+++ src/components/Badge/Badge.tsx
@@ -1,21 +1,51 @@
 import React from 'react';
-import BaseBadge, { BaseBadgeProps } from '../BaseBadge';
+import Flex, { FlexProps } from '../Flex';
 
-export type BadgeProps = BaseBadgeProps;
+export interface BadgeProps {
+  /** The color theme of the badge */
+  color: 'neutral' | 'grey' | 'blue' | 'pink' | 'red';
+}
 
 /** A badge is simply a visual label to accompany & characterize a certain text*/
-const Badge: React.FC<BadgeProps> = props => {
+const Badge: React.FC<BadgeProps> = ({ color, children, ...rest }) => {
+  const colorProps = (function() {
+    switch (color) {
+      case 'neutral':
+        return {
+          border: '1px solid',
+          borderColor: 'grey200',
+          bg: 'transparent',
+          color: 'grey300',
+        };
+      case 'grey':
+        return { bg: 'grey100', color: 'grey300' };
+      case 'blue':
+        return { bg: 'blue100', color: 'blue300' };
+      case 'pink':
+        return { bg: 'red100', color: 'red300' };
+      case 'red':
+      default:
+        return { bg: 'red300', color: 'white' };
+    }
+  })() as Partial<FlexProps>;
+
   return (
-    <BaseBadge
+    <Flex
       width="fit-content"
       minWidth="62px"
       textAlign="center"
       fontWeight="medium"
       borderRadius="small"
-      px={1}
-      py={1}
-      {...props}
-    />
+      align="center"
+      justify="center"
+      fontSize={0}
+      lineHeight={0}
+      p={1}
+      {...colorProps}
+      {...rest}
+    >
+      {children}
+    </Flex>
   );
 };
