--- src/components/BaseBadge/index.tsx
@@ -1,2 +1,0 @@
-export { default } from './BaseBadge';
-export * from './BaseBadge';
